+++ lib/calendar_data_timeline.dart
@@ -6,12 +6,17 @@ import 'package:calendar_data_timeline/models/export_all.dart';
 export 'package:calendar_data_timeline/models/export_all.dart';
 
 class CalendarDataTimeline extends StatefulWidget {
-  final List<CalendarContent> data;
-  final HeaderSettings settings;
-  final BodySettings bodySettings;
-
+  final List<CalendarContent>
+      data; //List of Instance CalendarContent which will be used as the in body
+  final HeaderSettings settings; //Instance of HeaderSettings
+  final BodySettings bodySettings; //Instance of BodySettings
+  final Color?
+      sundayColor; // Color choice for sunday color, leave it null if you want to highlight sunday as well.
   CalendarDataTimeline(
-      {required this.data, required this.settings, required this.bodySettings});
+      {required this.data,
+      required this.settings,
+      required this.bodySettings,
+      this.sundayColor});
 
   @override
   _CalendarDataTimelineState createState() => _CalendarDataTimelineState();
@@ -34,11 +39,13 @@ class _CalendarDataTimelineState extends State<CalendarDataTimeline> {
                 CalendarHeader(
                   itemLabel: widget.settings.label,
                   settings: widget.settings,
+                  sundayColor: widget.sundayColor,
                 ),
                 Expanded(
                   child: CalendarBody(
                     bodySettings: widget.bodySettings,
                     data: widget.data,
+                    sundayColor: widget.sundayColor,
                   ),
                 )
               ],