+++ lib/constant.dart
@@ -1,1 +1,1 @@
-const double percentage = 1.11;
+const double percentage = 1.11; // This is for cell size